added field - operator, more code in evaluate(), error()
added a new field that will store the current operator and will carry
over to a new operation when needed. Evaluate() now initialised the
required new fraction to return which will be initially set with the
values of the passed fraction as a parameter - this will ensure the
value of the previous calculation carries over too. Error () for now
prints an "error" message