else if expansion - operation with fractions
expanded the code to include identification of fractions in the string
input. It checks if there is an operator or not - if yes, then the
operation is executed. If not, then the value of the calculator becomes
the newly found fraction.
I made the fields public (temporarily) so that I can access them from my
FractionCalculatorTest class.
Also decided to remove isEmpty () boolean check and turn it
(operator==null).